Update README.md
This a Simple project that I have been working on while learning python programming. It is a sample code that accesses values on rows 2 to maximum row and column  3 and assigns a new value by multiplying it by 80%. The new values are then saved on a new cell on rows 2 to maximum row and column 4. The new values are then used to draw a Pie chart and a bar graph on a new excel file that is saved with a new name.
For more info email me on my email ##ocharodominic01@gmail.com##
Thank You.